EP-3251 #117 merged/ported various ImageCollectionClient tests to suite that runs aginst both 0.4.0 and 1.0.0 style cubes
made some fixtures more reusable in the process
