+++ tests/data/0.4.0/aggregate_zonal_path.json
@@ -19,7 +19,7 @@
     "loadcollection1": {
       "process_id": "load_collection",
       "arguments": {
-        "id": "SENTINEL2_FAPAR",
+        "id": "S2",
         "spatial_extent": null,
         "temporal_extent": null
       },
@@ -27,10 +27,10 @@
     },
     "readvector1": {
       "arguments": {
-        "filename":"/data/users/Public/vdboschj/EP-3025/GeometryCollection.geojson"
+        "filename": "/some/path/to/GeometryCollection.geojson"
       },
-      "process_id":"read_vector",
-      "result":false
+      "process_id": "read_vector",
+      "result": false
     },
     "aggregatepolygon1": {
       "process_id": "aggregate_polygon",
