+++ tests/data/0.4.0/aggregate_zonal_polygon.json
@@ -1,13 +1,5 @@
 {
   "process_graph": {
-    "loadcollection1": {
-      "process_id": "load_collection",
-      "arguments": {
-        "id": "SENTINEL2_FAPAR",
-        "spatial_extent": null,
-        "temporal_extent": null
-      }
-    },
     "filterbbox1": {
       "process_id": "filter_bbox",
       "arguments": {
@@ -21,7 +13,17 @@
           "south": 50,
           "crs": "EPSG:4326"
         }
-      }
+      },
+      "result": false
+    },
+    "loadcollection1": {
+      "process_id": "load_collection",
+      "arguments": {
+        "id": "S2",
+        "spatial_extent": null,
+        "temporal_extent": null
+      },
+      "result": false
     },
     "aggregatepolygon1": {
       "process_id": "aggregate_polygon",
@@ -63,14 +65,14 @@
           }
         },
         "reducer": {
-          "process_graph": {
-            "mean1": {
-              "process_id": "mean",
+          "callback": {
+            "unary": {
               "arguments": {
                 "data": {
                   "from_argument": "data"
                 }
               },
+              "process_id": "mean",
               "result": true
             }
           }
