+++ tests/data/0.4.0/pipe.json
@@ -1,0 +1,34 @@
+{
+  "loadcollection1": {
+    "process_id": "load_collection",
+    "arguments": {
+      "id": "S2",
+      "spatial_extent": null,
+      "temporal_extent": null
+    },
+    "result": false
+  },
+  "ndvi1": {
+    "process_id": "ndvi",
+    "arguments": {
+      "data": {
+        "from_node": "loadcollection1"
+      },
+      "name": "ndvi"
+    },
+    "result": false
+  },
+  "linearscalerange1": {
+    "process_id": "linear_scale_range",
+    "arguments": {
+      "x": {
+        "from_node": "ndvi1"
+      },
+      "inputMin": 0,
+      "inputMax": 1,
+      "outputMin": 0,
+      "outputMax": 100
+    },
+    "result": false
+  }
+}
