+++ tests/data/1.0.0/merge_ndvi_self.json
@@ -1,0 +1,38 @@
+{
+  "loadcollection1": {
+    "process_id": "load_collection",
+    "arguments": {
+      "id": "S2",
+      "spatial_extent": null,
+      "temporal_extent": null
+    }
+  },
+  "ndvi1": {
+    "process_id": "ndvi",
+    "arguments": {
+      "data": {
+        "from_node": "loadcollection1"
+      }
+    }
+  },
+  "loadcollection2": {
+    "process_id": "load_collection",
+    "arguments": {
+      "id": "S2",
+      "spatial_extent": null,
+      "temporal_extent": null
+    }
+  },
+  "mergecubes1": {
+    "process_id": "merge_cubes",
+    "arguments": {
+      "cube1": {
+        "from_node": "ndvi1"
+      },
+      "cube2": {
+        "from_node": "loadcollection2"
+      }
+    },
+    "result": true
+  }
+}
